feat: matricula integration

+++ solutions/FrontAngular/src/app/view/admin/curso/curso-search/curso-search.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit, ViewContainerRef } from '@angular/core';
 import { Curso } from 'src/app/model/curso';
 import { Router, ActivatedRoute } from '@angular/router';
 import { CursoService } from 'src/app/service/curso.service';
-import { TipoAcaoValues } from 'src/app/model/tipo-acao';
 import { MessagesService } from 'src/app/service/messages.service';
+import { TipoAcaoValues } from 'src/app/model/tipo-acao';
+import { TdDialogService } from '@covalent/core/dialogs';
 
 @Component({
   selector: 'app-curso-search',
@@ -19,13 +20,15 @@ export class CursoSearchComponent implements OnInit {
 
   /**
    * Construtor da classe
-   * @param router
-   * @param activatedRoute
+   * @param router 
+   * @param activatedRoute 
    */
   constructor(private router: Router,
             private activatedRoute: ActivatedRoute,
             private cursoService: CursoService,
-            private messageService: MessagesService) {
+            private messageService: MessagesService,
+            private _dialogService: TdDialogService,
+            private _viewContainerRef: ViewContainerRef) { 
   }
 
   /**
@@ -40,11 +43,12 @@ export class CursoSearchComponent implements OnInit {
    */
   navigateToNovo() {
     this.router.navigate(['cadastrar'], { relativeTo: this.activatedRoute });
+
   }
 
   /**
    * Método que redireciona para alterar, excluir ou visualizar curso
-   * @param evento
+   * @param evento 
    */
   navigateTo(evento) {
     console.log(evento.acaoRealizada);
@@ -54,11 +58,11 @@ export class CursoSearchComponent implements OnInit {
     }
     else if(evento.acaoRealizada == TipoAcaoValues[1]){
       this.router.navigate(['alterar/'+id], { relativeTo: this.activatedRoute });
-
+    
     } else if(evento.acaoRealizada == TipoAcaoValues[2]){
       this.remover(id);
     }
-
+    
   }
 
   /**
@@ -74,16 +78,37 @@ export class CursoSearchComponent implements OnInit {
       this.messageService.toastError(error.error.message);
     });
   }
-
+    
+  /**
+   * Método para remover um curso
+   */
   remover(id: number){
-    this.cursoService.remover(id).subscribe(dados => {
-      this.messageService.toastSuccess('Curso excluído com sucesso.');
-      this.listar();
-    },
-    (error: any) => {
-      console.log(error.error.message);
-      this.messageService.toastError(error.error.message);
+    this.openRemoverConfirm(id);
+  }
 
+  openRemoverConfirm(id: number): void {
+    this._dialogService.openConfirm({
+      message: 'Tem certeza que deseja excluir esse curso?',
+      disableClose: true, // defaults to false
+      viewContainerRef: this._viewContainerRef, //OPTIONAL
+      title: 'Excluir curso', //OPTIONAL, hides if not provided
+      cancelButton: 'Não', //OPTIONAL, defaults to 'CANCEL'
+      acceptButton: 'Sim', //OPTIONAL, defaults to 'ACCEPT'
+      width: '500px', //OPTIONAL, defaults to 400px
+    }).afterClosed().subscribe((accept: boolean) => {
+      if (accept) {
+        this.cursoService.remover(id).subscribe(dados => {
+          this.messageService.toastSuccess('Curso excluído com sucesso.');
+          this.listar();
+        },
+        (error: any) => {
+          console.log(error.error.message);
+          this.messageService.toastError(error.error.message);
+
+        });
+      } else {
+        // DO SOMETHING ELSE
+      }
     });
   }
 }
